Move calculating score into a different file
